fixes cycle issue
