+++ src/react/utils.js
@@ -10,15 +10,11 @@
 /* @flow */
 
 import { Realm, Effects } from "../realm.js";
-import {
-  AbruptCompletion,
-  JoinedAbruptCompletions,
-  PossiblyNormalCompletion,
-  ThrowCompletion,
-} from "../completions.js";
+import { AbruptCompletion, PossiblyNormalCompletion } from "../completions.js";
 import type { BabelNode, BabelNodeJSXIdentifier } from "babel-types";
 import {
   AbstractObjectValue,
+  AbstractValue,
   Value,
   NumberValue,
   ObjectValue,
@@ -32,20 +28,16 @@ import {
   BooleanValue,
 } from "../values/index.js";
 import { Generator } from "../utils/generator.js";
-import type { Descriptor, ReactHint, PropertyBinding, ReactComponentTreeConfig } from "../types";
+import type { Descriptor, ReactHint, PropertyBinding, ReactComponentTreeConfig } from "../types.js";
 import { Get, cloneDescriptor } from "../methods/index.js";
 import { computeBinary } from "../evaluators/BinaryExpression.js";
 import type { ReactSerializerState, AdditionalFunctionEffects, ReactEvaluatedNode } from "../serializer/types.js";
 import invariant from "../invariant.js";
-import { Create, Properties } from "../singletons.js";
+import { Create, Properties, To } from "../singletons.js";
 import traverse from "babel-traverse";
 import * as t from "babel-types";
 import type { BabelNodeStatement } from "babel-types";
 import { CompilerDiagnostic, FatalError } from "../errors.js";
-import { ReconcilerRenderBailOut } from "./errors.js";
-import { To } from "../singletons.js";
-import AbstractValue from "../values/AbstractValue";
-import { EnvironmentRecord } from "../environment.js";
 
 export type ReactSymbolTypes =
   | "react.element"
@@ -845,44 +837,4 @@ export function sanitizeReactElementForFirstRenderOnly(realm: Realm, reactElemen
     }
   }
   return reactElement;
-}
-
-function checkForSideEffects(effects: Effects, evaluatedNode: ReactEvaluatedNode): void {
-  let { modifiedBindings, result } = effects;
-  for (let [binding] of modifiedBindings) {
-    let env = binding.environment;
-
-    if (!(env instanceof EnvironmentRecord)) {
-      let bindings = env.bindings;
-      let name = binding.name;
-      // ensure that this binding was created outside these effects
-      // it's pure to mutate a binding that was created in a pure function
-      if (bindings[name] === undefined) {
-        throw new ReconcilerRenderBailOut(
-          `Failed to optimize React component tree for "${
-            evaluatedNode.name
-          }" due to side-effects from mutating the binding "${name}". Try wrapping side-effects in __sideEffect()`,
-          evaluatedNode
-        );
-      }
-    }
-  }
-  if (result instanceof PossiblyNormalCompletion || result instanceof JoinedAbruptCompletions) {
-    if (result.alternate instanceof ThrowCompletion || result.consequent instanceof ThrowCompletion) {
-      throw new ReconcilerRenderBailOut(
-        `Failed to optimize React component tree for "${
-          evaluatedNode.name
-        }" due to an exception thrown during render phase`,
-        evaluatedNode
-      );
-    }
-    checkForSideEffects(result.alternateEffects, evaluatedNode);
-    checkForSideEffects(result.consequentEffects, evaluatedNode);
-  }
-}
-
-export function throwIfRenderWasNotPure(effects: Effects, evaluatedNode: ReactEvaluatedNode): void {
-  // TODO: should we also check realm.savedEffects?
-  // ref: https://github.com/facebook/prepack/pull/1742
-  checkForSideEffects(effects, evaluatedNode);
 }
